requirements.txt
psycopg2
pyotp 
qrcode 
pillow

